Update Forking class, cast values from database into int where needed

+++ Blacklight/libraries/Forking.php
@@ -377,20 +377,19 @@ class Forking extends \fork_daemon
 
     /**
      * @return int
-     * @throws \Exception
      */
     private function safeBackfillMainMethod(): int
     {
         $this->register_child_run([0 => $this, 1 => 'safeBackfillChildWorker']);
 
-        $backfill_qty = Settings::settingValue('site.tmux.backfill_qty');
-        $backfill_order = Settings::settingValue('site.tmux.backfill_order');
-        $backfill_days = Settings::settingValue('site.tmux.backfill_days');
-        $maxmssgs = Settings::settingValue('maxmssgs');
-        $threads = Settings::settingValue('..backfillthreads');
+        $backfill_qty = (int) Settings::settingValue('site.tmux.backfill_qty');
+        $backfill_order = (int) Settings::settingValue('site.tmux.backfill_order');
+        $backfill_days = (int) Settings::settingValue('site.tmux.backfill_days');
+        $maxmssgs = (int) Settings::settingValue('maxmssgs');
+        $threads = (int) Settings::settingValue('..backfillthreads');
 
         $orderby = 'ORDER BY a.last_record ASC';
-        switch ((int) $backfill_order) {
+        switch ($backfill_order) {
             case 1:
                 $orderby = 'ORDER BY first_record_postdate DESC';
                 break;
@@ -413,10 +412,10 @@ class Forking extends \fork_daemon
         }
 
         $backfilldays = '';
-        if ((int) $backfill_days === 1) {
+        if ($backfill_days === 1) {
             $days = 'backfill_target';
             $backfilldays = now()->subDays((int) $days)->format('Y-m-d');
-        } elseif ((int) $backfill_days === 2) {
+        } elseif ($backfill_days === 2) {
             $backfilldays = now()->subDays(Carbon::createFromFormat('Y-m-d', Settings::settingValue('..safebackfilldate'))->diffInDays())->format('Y-m-d');
         }
 
@@ -481,8 +480,7 @@ class Forking extends \fork_daemon
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
     /**
-     * @return null|string
-     * @throws \Exception
+     * @return int
      */
     private function binariesMainMethod()
     {
@@ -494,7 +492,7 @@ class Forking extends \fork_daemon
             )
         );
 
-        return Settings::settingValue('..binarythreads');
+        return (int) Settings::settingValue('..binarythreads');
     }
 
     /**
@@ -518,9 +516,9 @@ class Forking extends \fork_daemon
     {
         $this->register_child_run([0 => $this, 1 => 'safeBinariesChildWorker']);
 
-        $maxheaders = Settings::settingValue('..max_headers_iteration') ?: 1000000;
-        $maxmssgs = Settings::settingValue('..maxmssgs');
-        $threads = Settings::settingValue('..binarythreads');
+        $maxheaders = (int) Settings::settingValue('..max_headers_iteration') ?: 1000000;
+        $maxmssgs = (int) Settings::settingValue('..maxmssgs');
+        $threads = (int) Settings::settingValue('..binarythreads');
 
         $groups = DB::select(
             '
@@ -586,8 +584,7 @@ class Forking extends \fork_daemon
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
     /**
-     * @return int|null|string
-     * @throws \Exception
+     * @return int
      */
     private function fixRelNamesMainMethod()
     {
@@ -655,8 +652,7 @@ class Forking extends \fork_daemon
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
     /**
-     * @return null|string
-     * @throws \Exception
+     * @return int
      */
     private function releasesMainMethod()
     {
@@ -750,7 +746,7 @@ class Forking extends \fork_daemon
     }
 
     /**
-     * @return int|null|string
+     * @return int
      * @throws \Exception
      */
     private function postProcessAddMainMethod()
@@ -823,7 +819,7 @@ class Forking extends \fork_daemon
                     $this->nfoQueryString
                 )
             );
-            $maxProcesses = Settings::settingValue('..nfothreads');
+            $maxProcesses = (int) Settings::settingValue('..nfothreads');
         }
 
         return $maxProcesses;
